Move actual api stuff to API module
