+++ graphql-api.cabal
@@ -32,6 +32,7 @@ library
     , text
   exposed-modules:
       GraphQL
+      GraphQL.API
       GraphQL.Application
       GraphQL.Muckaround
       GraphQL.Schema
